+++ mongokit/distinct_test.go
@@ -72,7 +72,7 @@ func TestDistinct(t *testing.T) {
 	// numbers
 	distinctTest(t, bsonkit.List{
 		bsonkit.Convert(bson.M{"a": int32(1)}),
-		bsonkit.Convert(bson.M{"a": int64(1)}),
+		bsonkit.Convert(bson.M{"a": 1}),
 		bsonkit.Convert(bson.M{"a": float64(1)}),
 		bsonkit.Convert(bson.M{"a": "1"}),
 	}, func(fn func(string, bson.A)) {
